Add port 2793 to redirect
Somehow this method get's called with an already pre-populated absolute
url, which is missing the 2793 port. It gets generated in another class,
that simply checks if the request arrived on a standard https port and
and then generates the url.
Strange thing is that not always the X-Forwarded=Port header is present,
so it cannot be used  consistently to override the port. At least i
wasn't able to find a pattern for the short time i was looking into it